removed dead references

+++ src/disp/glapp.cc
@@ -1274,7 +1274,7 @@ void glApp::displayMain()
  */
 void glApp::postRedisplay()
 {
-    for ( unsigned n = 1; n < views.size(); ++n )
+    for ( size_t n = 1; n < views.size(); ++n )
         if ( views[n].window() > 0 )
             glutPostWindowRedisplay(n);
 }
